task2/requirements.txt
# Core dependencies for Creative Automation Pipeline
openai==1.3.5
pillow==10.1.0
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.0

# Azure cloud storage
azure-storage-blob==12.19.0

# Image processing and computer vision
opencv-python==4.8.1.78
numpy==1.24.3

# Data handling and validation
pydantic==2.5.0
jsonschema==4.20.0

# Logging and monitoring
loguru==0.7.2

# File handling and utilities
pathlib2==2.3.7
typing-extensions==4.8.0

# Optional: Advanced image processing
scikit-image==0.22.0

# Testing (for development)
pytest==7.4.3
pytest-cov==4.1.0